requirements.txt
pandas==2.3.2
numpy==2.1.0
streamlit==1.49.1
google-generativeai==0.8.5
openpyxl==3.1.5
plotly==5.22.0